# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AckKind": {
        "description": "Worker acknowledgment outcomes accepted by the queue boundary.",
        "enum": [
          "success",
          "retryable_failure"
        ],
        "title": "AckKind",
        "type": "string"
      },
...
      "AnalysisOutput": {
        "additionalProperties": false,
        "description": "Only valid relevance, sentiment, and topic combinations.",
        "properties": {
          "relevance": {
            "title": "Relevance",
            "type": "boolean"
          },
          "sentiment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Sentiment"
              },
              {
                "type": "null"
              }
            ]
          },
          "topics": {
            "items": {
              "maxLength": 80,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 20,
            "title": "Topics",
            "type": "array"
          }
        },
        "required": [
          "relevance",
          "sentiment",
          "topics"
        ],
        "title": "AnalysisOutput",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Sentiment": {
        "description": "Sentiment variants allowed for valid relevant analyses.",
        "enum": [
          "positive",
          "neutral",
          "negative"
        ],
        "title": "Sentiment",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkerAckPayload": {
        "additionalProperties": false,
        "description": "Success output or retryable failure bound to the original lease tuple.",
        "properties": {
          "content_hash": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Content Hash",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "item_id": {
            "format": "uuid",
            "title": "Item Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/AckKind"
          },
          "lease_token": {
            "maxLength": 128,
            "minLength": 32,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Lease Token",
            "type": "string"
          },
          "model_version": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Model Version",
            "type": "string"
          },
          "output": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisOutput"
              },
              {
                "type": "null"
              }
            ]
          },
          "post_version_id": {
            "format": "uuid",
            "title": "Post Version Id",
            "type": "string"
          },
          "prompt_version": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Prompt Version",
            "type": "string"
          },
          "schema_version": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "item_id",
          "post_version_id",
          "content_hash",
          "prompt_version",
          "model_version",
          "schema_version",
          "lease_token",
          "kind"
        ],
        "title": "WorkerAckPayload",
        "type": "object"
      },
...
      "WorkerHeartbeatPayload": {
        "additionalProperties": false,
        "description": "CAS tuple required to extend an active analysis lease.",
        "properties": {
          "content_hash": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Content Hash",
            "type": "string"
          },
          "item_id": {
            "format": "uuid",
            "title": "Item Id",
            "type": "string"
          },
          "lease_token": {
            "maxLength": 128,
            "minLength": 32,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Lease Token",
            "type": "string"
          },
          "model_version": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Model Version",
            "type": "string"
          },
          "post_version_id": {
            "format": "uuid",
            "title": "Post Version Id",
            "type": "string"
          },
          "prompt_version": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Prompt Version",
            "type": "string"
          },
          "schema_version": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "item_id",
          "post_version_id",
          "content_hash",
          "prompt_version",
          "model_version",
          "schema_version",
          "lease_token"
        ],
        "title": "WorkerHeartbeatPayload",
        "type": "object"
      },
      "WorkerHeartbeatResult": {
        "additionalProperties": false,
        "description": "Server-owned expiry returned after a successful heartbeat.",
        "properties": {
          "lease_expires_at": {
            "format": "date-time",
            "title": "Lease Expires At",
            "type": "string"
          }
        },
        "required": [
          "lease_expires_at"
        ],
        "title": "WorkerHeartbeatResult",
        "type": "object"
      },
      "WorkerLeaseBlocked": {
        "additionalProperties": false,
        "description": "Capability-blocked response that cannot contain work or a lease token.",
        "properties": {
          "outcome": {
            "const": "blocked_capability",
            "default": "blocked_capability",
            "title": "Outcome",
            "type": "string"
          },
          "reason_codes": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Reason Codes",
            "type": "array"
          }
        },
        "required": [
          "reason_codes"
        ],
        "title": "WorkerLeaseBlocked",
        "type": "object"
      },
      "WorkerLeaseEmpty": {
        "additionalProperties": false,
        "description": "Honest empty queue response without synthetic analysis work.",
        "properties": {
          "outcome": {
            "const": "empty",
            "default": "empty",
            "title": "Outcome",
            "type": "string"
          },
          "retry_after_seconds": {
            "maximum": 600.0,
            "minimum": 1.0,
            "title": "Retry After Seconds",
            "type": "integer"
          }
        },
        "required": [
          "retry_after_seconds"
        ],
        "title": "WorkerLeaseEmpty",
        "type": "object"
      },
      "WorkerLeaseGranted": {
        "additionalProperties": false,
        "description": "One version-bound work item leased to the worker.",
        "properties": {
          "item": {
            "$ref": "#/components/schemas/WorkerLeaseItem"
          },
          "outcome": {
            "const": "leased",
            "default": "leased",
            "title": "Outcome",
            "type": "string"
          }
        },
        "required": [
          "item"
        ],
        "title": "WorkerLeaseGranted",
        "type": "object"
      },
      "WorkerLeaseItem": {
        "additionalProperties": false,
        "description": "Author-free analysis input and one response-only raw lease token.",
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "content_hash": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Content Hash",
            "type": "string"
          },
          "item_id": {
            "format": "uuid",
            "title": "Item Id",
            "type": "string"
          },
          "language": {
            "enum": [
              "ko",
              "en"
            ],
            "title": "Language",
            "type": "string"
          },
          "lease_expires_at": {
            "format": "date-time",
            "title": "Lease Expires At",
            "type": "string"
          },
          "lease_token": {
            "maxLength": 128,
            "minLength": 32,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Lease Token",
            "type": "string"
          },
          "model_version": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Model Version",
            "type": "string"
          },
          "post_version_id": {
            "format": "uuid",
            "title": "Post Version Id",
            "type": "string"
          },
          "prompt_version": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Prompt Version",
            "type": "string"
          },
          "schema_version": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Schema Version",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "item_id",
          "post_version_id",
          "content_hash",
          "prompt_version",
          "model_version",
          "schema_version",
          "lease_token",
          "title",
          "body",
          "language",
          "lease_expires_at"
        ],
        "title": "WorkerLeaseItem",
        "type": "object"
      },
      "WorkerLeaseRequest": {
        "additionalProperties": false,
        "description": "Current capability proof requested by the authenticated worker.",
        "properties": {
          "capability_proof_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Capability Proof Id",
            "type": "string"
          }
        },
        "required": [
          "capability_proof_id"
        ],
        "title": "WorkerLeaseRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Prediction-market community monitor API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/worker/ack": {
      "post": {
        "description": "Acknowledge an exact active lease after worker:ack authorization.",
        "operationId": "ack_v1_worker_ack_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkerAckPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ack",
        "tags": [
          "worker"
        ]
      }
    },
    "/v1/worker/heartbeat": {
      "post": {
        "description": "Extend an exact active lease after worker:heartbeat authorization.",
        "operationId": "heartbeat_v1_worker_heartbeat_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkerHeartbeatPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerHeartbeatResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Heartbeat",
        "tags": [
          "worker"
        ]
      }
    },
    "/v1/worker/lease": {
      "post": {
        "description": "Lease work only after worker:lease authorization.",
        "operationId": "lease_v1_worker_lease_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkerLeaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "discriminator": {
                    "mapping": {
                      "blocked_capability": "#/components/schemas/WorkerLeaseBlocked",
                      "empty": "#/components/schemas/WorkerLeaseEmpty",
                      "leased": "#/components/schemas/WorkerLeaseGranted"
                    },
                    "propertyName": "outcome"
                  },
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/WorkerLeaseGranted"
                    },
                    {
                      "$ref": "#/components/schemas/WorkerLeaseBlocked"
                    },
                    {
                      "$ref": "#/components/schemas/WorkerLeaseEmpty"
                    }
                  ],
                  "title": "Response Lease V1 Worker Lease Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lease",
        "tags": [
          "worker"
        ]
      }
    }
  }
}